refacto / add comments for :arrow_right: Database/Todos

+++ Database/Todos/todos_deleteALL_database.php
@@ -3,13 +3,11 @@ session_start();
 include('../connection_database.php');
 include('../../login_database.php');
 
-$deleteALL_todos = $bdd->prepare('DELETE FROM create_todos WHERE id_user = :id_user');
-$deleteALL_todos->bindParam(':id_user', $_SESSION['id_user']);
-$deleteALL_todos->execute();
-
-if($deleteALL_todos){
-    header('Location: ../../layouts/todos.php');
-}else{
-    echo "Error deleting todos";
+// Delete every todos
+if($_SESSION['id_user']){
+    $deleteALL_todos = $bdd->prepare('DELETE FROM create_todos WHERE id_user = :id_user');
+    $deleteALL_todos->bindParam(':id_user', $_SESSION['id_user']);
+    $deleteALL_todos->execute();
 }
+header('Location: ../../layouts/todos.php');
 ?>
